requirements.txt
wemake-python-styleguide==0.19.2
albumentations==1.3.1
omegaconf==2.3.0
opencv-python==4.8.0.74
iterative_stratification==0.1.7
scikit-multilearn==0.2.0
numpy==1.24.4
pandas==2.0.3
pydantic==1.10.11
pytorch_lightning[extra]==2.0.5
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
torchmetrics== 1.0.1
clearml==1.11.1
pre-commit==3.8.0
openvino==2024.1.0
segmentation-models-pytorch==0.3.3
onnxsim==0.4.36
onnxruntime==1.18.1
nltk==3.8.1
